Added a partitioner, to convert the processed csv.gz files that are partitioned by time (and sorted internally by time) into files partitioned geographically, producing files with similar numbers of records in each, so busy areas have files covering a small area, and out in the suburbs where there are fewer routes and more area, there are files covering large regions.
Started work on resurrecting the code for fitting points
to a path.

Started on a program for recovering the VehicleLocationReport
from a saved .xml file (response from NextBus), with the
aim of being able to produce raw csv.gz files, with multiple
records for a single vehicle report in the case that we received
multiple such reports from NextBus, which we can then aggregate
into processed csv.gz files.